docs: update cookbook continue after errors example
