+++ docs/current_docs/cookbook/cookbook.mdx
@@ -1155,7 +1155,7 @@ The caveat with this approach is that forcing a zero exit code on a failure cach
 </TabItem>
 <TabItem value="TypeScript">
 
-```typescript file=./snippets/continue-after-errors/typescript/index.ts
+```typescript file=./snippets/continue-after-errors/typescript/src/index.ts
 ```
 
 </TabItem>
