+++ docs/current_docs/cookbook/snippets/continue-after-errors/go/dagger.json
@@ -2,5 +2,5 @@
   "name": "my-module",
   "sdk": "go",
   "source": ".",
-  "engineVersion": "v0.12.0"
+  "engineVersion": "v0.13.6"
 }
