+++ docs/current_docs/cookbook/snippets/continue-after-errors/go/main.go
@@ -12,15 +12,15 @@ type MyModule struct{}
 var script = `#!/bin/sh
 echo "Test Suite"
 echo "=========="
-echo "Test 1: PASS" >> report.txt
-echo "Test 2: FAIL" >> report.txt
-echo "Test 3: PASS" >> report.txt
+echo "Test 1: PASS" | tee -a report.txt
+echo "Test 2: FAIL" | tee -a report.txt
+echo "Test 3: PASS" | tee -a report.txt
 exit 1
 `
 
 type TestResult struct {
 	Report   *dagger.File
-	ExitCode string
+	ExitCode int
 }
 
 // Handle errors
@@ -29,9 +29,9 @@ func (m *MyModule) Test(ctx context.Context) (*TestResult, error) {
 		Container().
 		From("alpine").
 		// add script with execution permission to simulate a testing tool
-		WithNewFile("run-tests", script, dagger.ContainerWithNewFileOpts{Permissions: 0o750}).
-		// if the exit code isn't needed: "run-tests; true"
-		WithExec([]string{"sh", "-c", "/run-tests; echo -n $? > /exit_code"}).
+		WithNewFile("/run-tests", script, dagger.ContainerWithNewFileOpts{Permissions: 0o750}).
+		// run-tests but allow any return code
+		WithExec([]string{"/run-tests"}, dagger.ContainerWithExecOpts{Expect: dagger.Any}).
 		// the result of `sync` is the container, which allows continued chaining
 		Sync(ctx)
 	if err != nil {
@@ -42,7 +42,7 @@ func (m *MyModule) Test(ctx context.Context) (*TestResult, error) {
 	report := ctr.File("report.txt")
 
 	// use the saved exit code to determine if the tests passed.
-	exitCode, err := ctr.File("/exit_code").Contents(ctx)
+	exitCode, err := ctr.ExitCode(ctx)
 	if err != nil {
 		// exit code not found
 		return nil, fmt.Errorf("get exit code: %w", err)
