+++ docs/current_docs/cookbook/snippets/continue-after-errors/python/dagger.json
@@ -1,6 +1,6 @@
 {
   "name": "my-module",
   "sdk": "python",
-  "source": "dagger",
-  "engineVersion": "v0.11.5"
+  "source": ".",
+  "engineVersion": "v0.13.6"
 }
