+++ docs/current_docs/cookbook/snippets/continue-after-errors/python/main.py
@@ -4,9 +4,9 @@ from dagger import DaggerError, dag, field, function, object_type
 SCRIPT = """#!/bin/sh
 echo "Test Suite"
 echo "=========="
-echo "Test 1: PASS" >> report.txt
-echo "Test 2: FAIL" >> report.txt
-echo "Test 3: PASS" >> report.txt
+echo "Test 1: PASS" | tee -a report.txt
+echo "Test 2: FAIL" | tee -a report.txt
+echo "Test 3: PASS" | tee -a report.txt
 exit 1
 """
 
@@ -14,7 +14,7 @@ exit 1
 @object_type
 class TestResult:
     report: dagger.File = field()
-    exit_code: str = field()
+    exit_code: int = field()
 
 
 @object_type
@@ -27,9 +27,9 @@ class MyModule:
                 dag.container()
                 .from_("alpine")
                 # add script with execution permission to simulate a testing tool.
-                .with_new_file("run-tests", SCRIPT, permissions=0o750)
-                # if the exit code isn't needed: "run-tests; true"
-                .with_exec(["sh", "-c", "/run-tests; echo -n $? > /exit_code"])
+                .with_new_file("/run-tests", SCRIPT, permissions=0o750)
+                # run-tests but allow any return code
+                .with_exec(["/run-tests"], expect=dagger.ReturnType.ANY)
                 # the result of `sync` is the container, which allows continued chaining
                 .sync()
             )
@@ -38,7 +38,7 @@ class MyModule:
             report = ctr.file("report.txt")
 
             # use the saved exit code to determine if the tests passed.
-            exit_code = await ctr.file("exit_code").contents()
+            exit_code = await ctr.exit_code()
 
             return TestResult(report=report, exit_code=exit_code)
         except DaggerError as e:
