+++ docs/current_docs/cookbook/snippets/continue-after-errors/typescript/dagger.json
@@ -1,6 +1,6 @@
 {
   "name": "my-module",
   "sdk": "typescript",
-  "source": "dagger",
-  "engineVersion": "v0.11.5"
+  "source": ".",
+  "engineVersion": "v0.13.7"
 }
